moved image function to utils
